Switched to rendering via the Java2D graphics API. Finished up first pass. Everything mostly works and I can visualize my projects and end up with a useful set of printouts.
git-svn-id: https://samskivert.googlecode.com/svn/trunk@186 6335cc39-0255-0410-8fd6-9bcaacd3b74c

+++ projects/viztool/src/java/com/samskivert/viztool/layout/CascadingConstants.java
@@ -1,5 +1,5 @@
 //
-// $Id: CascadingConstants.java,v 1.1 2001/07/13 23:25:13 mdb Exp $
+// $Id: CascadingConstants.java,v 1.2 2001/07/17 01:54:19 mdb Exp $
 
 package com.samskivert.viztool.viz;
 
@@ -14,10 +14,10 @@ public interface CascadingConstants
     /**
      * The number of points surrounding the name of the chain.
      */
-    public static final int HEADER_BORDER = 3;
+    public static final double HEADER_BORDER = 3;
 
     /**
      * The number of points of spacing between each child chain.
      */
-    public static final int GAP = 4;
+    public static final double GAP = 4;
 }
